fix: added support for fallen-list's skeleton to be shown
