+++ app/all-fallen/page.jsx
@@ -1,17 +1,11 @@
 import { Suspense } from "react";
 import styles from "./page.module.scss";
-import { connectToDB } from "@/server/connect";
 import FallenList from "../components/FallenList";
 import SearchForm from "../components/SearchForm";
-import Pagination from "../components/Pagination";
 import TitleDivider from "../components/TitleDivider";
 import { metadata as layoutMetadata } from "../layout";
 import PopularHobbies from "../components/PopularHobbies";
 import FallenListSkeleton from "../components/Skeletons/FallenListSkeleton";
-import {
-  getAllFallen,
-  getFilteredFallen,
-} from "@/server/service/fallen.service";
 
 export const metadata = {
   title: "כל הנופלים",
@@ -28,24 +22,13 @@ export const metadata = {
 };
 
 export default async function AllFallenPage({ searchParams }) {
-  await connectToDB();
-
   // Get params
   const params = await searchParams;
   const query = params.q || "";
   const currentPage = Number(params.page) || 1;
 
-  // Configure pagination settings
+  // Configure pagination's limit
   const limit = 10;
-  const skip = (currentPage - 1) * limit;
-
-  // Fetch fallen data based on query and pagination settings
-  const { data, total } = query
-    ? await getFilteredFallen(query, limit, skip)
-    : await getAllFallen(limit, skip);
-
-  // Calculate total pages based on the total number of fallen and the limit per page
-  const totalPages = Math.ceil(total / limit);
 
   return (
     <>
@@ -57,18 +40,12 @@ export default async function AllFallenPage({ searchParams }) {
       {/* Hobbies filter section */}
       <PopularHobbies containerType="tag" />
 
-      {/* Fallen list section */}
+      {/* Fallen list and pagination section */}
       <div className={styles.itemsContainer}>
         <Suspense fallback={<FallenListSkeleton limit={limit} />}>
-          <FallenList fallen={data} />
+          <FallenList query={query} currentPage={currentPage} limit={limit} />
         </Suspense>
       </div>
-
-      {/* Pagination section */}
-      <Pagination
-        currentPage={total === 0 ? total : currentPage}
-        totalPages={totalPages}
-      />
     </>
   );
 }
