+++ app/components/Pagination/style.module.scss
@@ -1,6 +1,7 @@
 @use '@/styles/_vars.scss' as *;
 
 .paginationContainer {
+    width: 100%;
     display: flex;
     justify-content: space-around;
     align-items: center;
